click: fix camera failure after zoom usage
libcamera was changing the image resolution without updating
mRawWidth and mRawHeight, and this caused initRaw to take pictures
with the previous resolution. If that resolution was unsuitable
because of zoom level, it would cause the camera to fail until
it was restarted.

This patch updates both variables to reflect the modified
resolution so initRaw always takes pictures with the correct
width and height.

Change-Id: I33e8445a65ac667a030b35b103f86be46381abc4